+++ ecs/query_gen.go
@@ -167,6 +167,7 @@ type Query1[A any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
 	relations  []relationID
@@ -178,6 +179,9 @@ type Query1[A any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query1[A]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -228,6 +232,7 @@ func (q *Query1[A]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
 	q.world.unlockSafe(q.lock)
@@ -294,9 +299,9 @@ func (q *Query1[A]) nextTable(tables []tableID) bool {
 func (q *Query1[A]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -310,8 +315,10 @@ type Query2[A any, B any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
 	relations  []relationID
@@ -323,6 +330,9 @@ type Query2[A any, B any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query2[A, B]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -369,8 +379,10 @@ func (q *Query2[A, B]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
 	q.world.unlockSafe(q.lock)
@@ -437,12 +449,12 @@ func (q *Query2[A, B]) nextTable(tables []tableID) bool {
 func (q *Query2[A, B]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -458,10 +470,13 @@ type Query3[A any, B any, C any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
 	relations  []relationID
@@ -473,6 +488,9 @@ type Query3[A any, B any, C any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query3[A, B, C]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -523,10 +541,13 @@ func (q *Query3[A, B, C]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
 	q.world.unlockSafe(q.lock)
@@ -593,15 +614,15 @@ func (q *Query3[A, B, C]) nextTable(tables []tableID) bool {
 func (q *Query3[A, B, C]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -617,12 +638,16 @@ type Query4[A any, B any, C any, D any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
+	columnD    *column
 	columnPtrD unsafe.Pointer
 	itemSizeD  uintptr
 	relations  []relationID
@@ -634,6 +659,9 @@ type Query4[A any, B any, C any, D any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query4[A, B, C, D]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -684,12 +712,16 @@ func (q *Query4[A, B, C, D]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
+	q.columnD = nil
 	q.columnPtrD = unsafe.Pointer(nilDummy)
 	q.itemSizeD = 0
 	q.world.unlockSafe(q.lock)
@@ -756,18 +788,18 @@ func (q *Query4[A, B, C, D]) nextTable(tables []tableID) bool {
 func (q *Query4[A, B, C, D]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
-	columnD := q.components[3].columns[q.table.id]
-	q.columnPtrD = columnD.pointer
-	q.itemSizeD = columnD.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
+	q.columnD = q.components[3].columns[q.table.id]
+	q.columnPtrD = q.columnD.pointer
+	q.itemSizeD = q.columnD.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -783,14 +815,19 @@ type Query5[A any, B any, C any, D any, E any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
+	columnD    *column
 	columnPtrD unsafe.Pointer
 	itemSizeD  uintptr
+	columnE    *column
 	columnPtrE unsafe.Pointer
 	itemSizeE  uintptr
 	relations  []relationID
@@ -802,6 +839,9 @@ type Query5[A any, B any, C any, D any, E any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query5[A, B, C, D, E]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -852,14 +892,19 @@ func (q *Query5[A, B, C, D, E]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
+	q.columnD = nil
 	q.columnPtrD = unsafe.Pointer(nilDummy)
 	q.itemSizeD = 0
+	q.columnE = nil
 	q.columnPtrE = unsafe.Pointer(nilDummy)
 	q.itemSizeE = 0
 	q.world.unlockSafe(q.lock)
@@ -926,21 +971,21 @@ func (q *Query5[A, B, C, D, E]) nextTable(tables []tableID) bool {
 func (q *Query5[A, B, C, D, E]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
-	columnD := q.components[3].columns[q.table.id]
-	q.columnPtrD = columnD.pointer
-	q.itemSizeD = columnD.itemSize
-	columnE := q.components[4].columns[q.table.id]
-	q.columnPtrE = columnE.pointer
-	q.itemSizeE = columnE.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
+	q.columnD = q.components[3].columns[q.table.id]
+	q.columnPtrD = q.columnD.pointer
+	q.itemSizeD = q.columnD.itemSize
+	q.columnE = q.components[4].columns[q.table.id]
+	q.columnPtrE = q.columnE.pointer
+	q.itemSizeE = q.columnE.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -956,16 +1001,22 @@ type Query6[A any, B any, C any, D any, E any, F any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
+	columnD    *column
 	columnPtrD unsafe.Pointer
 	itemSizeD  uintptr
+	columnE    *column
 	columnPtrE unsafe.Pointer
 	itemSizeE  uintptr
+	columnF    *column
 	columnPtrF unsafe.Pointer
 	itemSizeF  uintptr
 	relations  []relationID
@@ -977,6 +1028,9 @@ type Query6[A any, B any, C any, D any, E any, F any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query6[A, B, C, D, E, F]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -1027,16 +1081,22 @@ func (q *Query6[A, B, C, D, E, F]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
+	q.columnD = nil
 	q.columnPtrD = unsafe.Pointer(nilDummy)
 	q.itemSizeD = 0
+	q.columnE = nil
 	q.columnPtrE = unsafe.Pointer(nilDummy)
 	q.itemSizeE = 0
+	q.columnF = nil
 	q.columnPtrF = unsafe.Pointer(nilDummy)
 	q.itemSizeF = 0
 	q.world.unlockSafe(q.lock)
@@ -1103,24 +1163,24 @@ func (q *Query6[A, B, C, D, E, F]) nextTable(tables []tableID) bool {
 func (q *Query6[A, B, C, D, E, F]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
-	columnD := q.components[3].columns[q.table.id]
-	q.columnPtrD = columnD.pointer
-	q.itemSizeD = columnD.itemSize
-	columnE := q.components[4].columns[q.table.id]
-	q.columnPtrE = columnE.pointer
-	q.itemSizeE = columnE.itemSize
-	columnF := q.components[5].columns[q.table.id]
-	q.columnPtrF = columnF.pointer
-	q.itemSizeF = columnF.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
+	q.columnD = q.components[3].columns[q.table.id]
+	q.columnPtrD = q.columnD.pointer
+	q.itemSizeD = q.columnD.itemSize
+	q.columnE = q.components[4].columns[q.table.id]
+	q.columnPtrE = q.columnE.pointer
+	q.itemSizeE = q.columnE.itemSize
+	q.columnF = q.components[5].columns[q.table.id]
+	q.columnPtrF = q.columnF.pointer
+	q.itemSizeF = q.columnF.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -1136,18 +1196,25 @@ type Query7[A any, B any, C any, D any, E any, F any, G any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
+	columnD    *column
 	columnPtrD unsafe.Pointer
 	itemSizeD  uintptr
+	columnE    *column
 	columnPtrE unsafe.Pointer
 	itemSizeE  uintptr
+	columnF    *column
 	columnPtrF unsafe.Pointer
 	itemSizeF  uintptr
+	columnG    *column
 	columnPtrG unsafe.Pointer
 	itemSizeG  uintptr
 	relations  []relationID
@@ -1159,6 +1226,9 @@ type Query7[A any, B any, C any, D any, E any, F any, G any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query7[A, B, C, D, E, F, G]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -1209,18 +1279,25 @@ func (q *Query7[A, B, C, D, E, F, G]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
+	q.columnD = nil
 	q.columnPtrD = unsafe.Pointer(nilDummy)
 	q.itemSizeD = 0
+	q.columnE = nil
 	q.columnPtrE = unsafe.Pointer(nilDummy)
 	q.itemSizeE = 0
+	q.columnF = nil
 	q.columnPtrF = unsafe.Pointer(nilDummy)
 	q.itemSizeF = 0
+	q.columnG = nil
 	q.columnPtrG = unsafe.Pointer(nilDummy)
 	q.itemSizeG = 0
 	q.world.unlockSafe(q.lock)
@@ -1287,27 +1364,27 @@ func (q *Query7[A, B, C, D, E, F, G]) nextTable(tables []tableID) bool {
 func (q *Query7[A, B, C, D, E, F, G]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
-	columnD := q.components[3].columns[q.table.id]
-	q.columnPtrD = columnD.pointer
-	q.itemSizeD = columnD.itemSize
-	columnE := q.components[4].columns[q.table.id]
-	q.columnPtrE = columnE.pointer
-	q.itemSizeE = columnE.itemSize
-	columnF := q.components[5].columns[q.table.id]
-	q.columnPtrF = columnF.pointer
-	q.itemSizeF = columnF.itemSize
-	columnG := q.components[6].columns[q.table.id]
-	q.columnPtrG = columnG.pointer
-	q.itemSizeG = columnG.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
+	q.columnD = q.components[3].columns[q.table.id]
+	q.columnPtrD = q.columnD.pointer
+	q.itemSizeD = q.columnD.itemSize
+	q.columnE = q.components[4].columns[q.table.id]
+	q.columnPtrE = q.columnE.pointer
+	q.itemSizeE = q.columnE.itemSize
+	q.columnF = q.components[5].columns[q.table.id]
+	q.columnPtrF = q.columnF.pointer
+	q.itemSizeF = q.columnF.itemSize
+	q.columnG = q.components[6].columns[q.table.id]
+	q.columnPtrG = q.columnG.pointer
+	q.itemSizeG = q.columnG.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }
@@ -1323,20 +1400,28 @@ type Query8[A any, B any, C any, D any, E any, F any, G any, H any] struct {
 	filter     *filter
 	table      *table
 	cache      *cacheEntry
+	columnA    *column
 	columnPtrA unsafe.Pointer
 	itemSizeA  uintptr
+	columnB    *column
 	columnPtrB unsafe.Pointer
 	itemSizeB  uintptr
+	columnC    *column
 	columnPtrC unsafe.Pointer
 	itemSizeC  uintptr
+	columnD    *column
 	columnPtrD unsafe.Pointer
 	itemSizeD  uintptr
+	columnE    *column
 	columnPtrE unsafe.Pointer
 	itemSizeE  uintptr
+	columnF    *column
 	columnPtrF unsafe.Pointer
 	itemSizeF  uintptr
+	columnG    *column
 	columnPtrG unsafe.Pointer
 	itemSizeG  uintptr
+	columnH    *column
 	columnPtrH unsafe.Pointer
 	itemSizeH  uintptr
 	relations  []relationID
@@ -1348,6 +1433,9 @@ type Query8[A any, B any, C any, D any, E any, F any, G any, H any] struct {
 }
 
 // GetRelation returns the entity relation target of the component at the given index.
+//
+// Can be used for the current entity in entity-based iteration,
+// as well as for the entire current table in table-based iteration.
 func (q *Query8[A, B, C, D, E, F, G, H]) GetRelation(index int) Entity {
 	return q.components[index].columns[q.table.id].target
 }
@@ -1398,20 +1486,28 @@ func (q *Query8[A, B, C, D, E, F, G, H]) Close() {
 	q.tables = nil
 	q.table = nil
 	q.cache = nil
+	q.columnA = nil
 	q.columnPtrA = unsafe.Pointer(nilDummy)
 	q.itemSizeA = 0
+	q.columnB = nil
 	q.columnPtrB = unsafe.Pointer(nilDummy)
 	q.itemSizeB = 0
+	q.columnC = nil
 	q.columnPtrC = unsafe.Pointer(nilDummy)
 	q.itemSizeC = 0
+	q.columnD = nil
 	q.columnPtrD = unsafe.Pointer(nilDummy)
 	q.itemSizeD = 0
+	q.columnE = nil
 	q.columnPtrE = unsafe.Pointer(nilDummy)
 	q.itemSizeE = 0
+	q.columnF = nil
 	q.columnPtrF = unsafe.Pointer(nilDummy)
 	q.itemSizeF = 0
+	q.columnG = nil
 	q.columnPtrG = unsafe.Pointer(nilDummy)
 	q.itemSizeG = 0
+	q.columnH = nil
 	q.columnPtrH = unsafe.Pointer(nilDummy)
 	q.itemSizeH = 0
 	q.world.unlockSafe(q.lock)
@@ -1478,30 +1574,30 @@ func (q *Query8[A, B, C, D, E, F, G, H]) nextTable(tables []tableID) bool {
 func (q *Query8[A, B, C, D, E, F, G, H]) setTable(index int32, table *table) {
 	q.cursor.table = index
 	q.table = table
-	columnA := q.components[0].columns[q.table.id]
-	q.columnPtrA = columnA.pointer
-	q.itemSizeA = columnA.itemSize
-	columnB := q.components[1].columns[q.table.id]
-	q.columnPtrB = columnB.pointer
-	q.itemSizeB = columnB.itemSize
-	columnC := q.components[2].columns[q.table.id]
-	q.columnPtrC = columnC.pointer
-	q.itemSizeC = columnC.itemSize
-	columnD := q.components[3].columns[q.table.id]
-	q.columnPtrD = columnD.pointer
-	q.itemSizeD = columnD.itemSize
-	columnE := q.components[4].columns[q.table.id]
-	q.columnPtrE = columnE.pointer
-	q.itemSizeE = columnE.itemSize
-	columnF := q.components[5].columns[q.table.id]
-	q.columnPtrF = columnF.pointer
-	q.itemSizeF = columnF.itemSize
-	columnG := q.components[6].columns[q.table.id]
-	q.columnPtrG = columnG.pointer
-	q.itemSizeG = columnG.itemSize
-	columnH := q.components[7].columns[q.table.id]
-	q.columnPtrH = columnH.pointer
-	q.itemSizeH = columnH.itemSize
+	q.columnA = q.components[0].columns[q.table.id]
+	q.columnPtrA = q.columnA.pointer
+	q.itemSizeA = q.columnA.itemSize
+	q.columnB = q.components[1].columns[q.table.id]
+	q.columnPtrB = q.columnB.pointer
+	q.itemSizeB = q.columnB.itemSize
+	q.columnC = q.components[2].columns[q.table.id]
+	q.columnPtrC = q.columnC.pointer
+	q.itemSizeC = q.columnC.itemSize
+	q.columnD = q.components[3].columns[q.table.id]
+	q.columnPtrD = q.columnD.pointer
+	q.itemSizeD = q.columnD.itemSize
+	q.columnE = q.components[4].columns[q.table.id]
+	q.columnPtrE = q.columnE.pointer
+	q.itemSizeE = q.columnE.itemSize
+	q.columnF = q.components[5].columns[q.table.id]
+	q.columnPtrF = q.columnF.pointer
+	q.itemSizeF = q.columnF.itemSize
+	q.columnG = q.components[6].columns[q.table.id]
+	q.columnPtrG = q.columnG.pointer
+	q.itemSizeG = q.columnG.itemSize
+	q.columnH = q.components[7].columns[q.table.id]
+	q.columnPtrH = q.columnH.pointer
+	q.itemSizeH = q.columnH.itemSize
 	q.cursor.index = 0
 	q.cursor.maxIndex = int64(q.table.len - 1)
 }