+++ ecs/query_nodebug_gen.go
@@ -7,6 +7,8 @@ package ecs
 import "unsafe"
 
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query0.NextTable].
 func (q *Query0) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -15,12 +17,30 @@ func (q *Query0) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query0.Next].
+func (q *Query0) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query0.Next].
 func (q *Query0) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query0.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query0) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query1.NextTable].
 func (q *Query1[A]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -29,12 +49,29 @@ func (q *Query1[A]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query1.Next].
+func (q *Query1[A]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query1.Next].
 func (q *Query1[A]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query1.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query1[A]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query1.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query1[A]) Get() *A {
@@ -42,7 +79,15 @@ func (q *Query1[A]) Get() *A {
 	return (*A)(unsafe.Add(q.columnPtrA, index*q.itemSizeA))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query1.NextTable].
+func (q *Query1[A]) GetColumns() []A {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query2.NextTable].
 func (q *Query2[A, B]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -51,12 +96,29 @@ func (q *Query2[A, B]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query2.Next].
+func (q *Query2[A, B]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query2.Next].
 func (q *Query2[A, B]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query2.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query2[A, B]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query2.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query2[A, B]) Get() (*A, *B) {
@@ -65,7 +127,16 @@ func (q *Query2[A, B]) Get() (*A, *B) {
 		(*B)(unsafe.Add(q.columnPtrB, index*q.itemSizeB))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query2.NextTable].
+func (q *Query2[A, B]) GetColumns() ([]A, []B) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query3.NextTable].
 func (q *Query3[A, B, C]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -74,12 +145,29 @@ func (q *Query3[A, B, C]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query3.Next].
+func (q *Query3[A, B, C]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query3.Next].
 func (q *Query3[A, B, C]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query3.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query3[A, B, C]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query3.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query3[A, B, C]) Get() (*A, *B, *C) {
@@ -89,7 +177,17 @@ func (q *Query3[A, B, C]) Get() (*A, *B, *C) {
 		(*C)(unsafe.Add(q.columnPtrC, index*q.itemSizeC))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query3.NextTable].
+func (q *Query3[A, B, C]) GetColumns() ([]A, []B, []C) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query4.NextTable].
 func (q *Query4[A, B, C, D]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -98,12 +196,29 @@ func (q *Query4[A, B, C, D]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query4.Next].
+func (q *Query4[A, B, C, D]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query4.Next].
 func (q *Query4[A, B, C, D]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query4.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query4[A, B, C, D]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query4.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query4[A, B, C, D]) Get() (*A, *B, *C, *D) {
@@ -114,7 +229,18 @@ func (q *Query4[A, B, C, D]) Get() (*A, *B, *C, *D) {
 		(*D)(unsafe.Add(q.columnPtrD, index*q.itemSizeD))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query4.NextTable].
+func (q *Query4[A, B, C, D]) GetColumns() ([]A, []B, []C, []D) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len],
+		q.columnD.data.Interface().([]D)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query5.NextTable].
 func (q *Query5[A, B, C, D, E]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -123,12 +249,29 @@ func (q *Query5[A, B, C, D, E]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query5.Next].
+func (q *Query5[A, B, C, D, E]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query5.Next].
 func (q *Query5[A, B, C, D, E]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query5.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query5[A, B, C, D, E]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query5.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query5[A, B, C, D, E]) Get() (*A, *B, *C, *D, *E) {
@@ -140,7 +283,19 @@ func (q *Query5[A, B, C, D, E]) Get() (*A, *B, *C, *D, *E) {
 		(*E)(unsafe.Add(q.columnPtrE, index*q.itemSizeE))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query5.NextTable].
+func (q *Query5[A, B, C, D, E]) GetColumns() ([]A, []B, []C, []D, []E) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len],
+		q.columnD.data.Interface().([]D)[:q.table.len:q.table.len],
+		q.columnE.data.Interface().([]E)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query6.NextTable].
 func (q *Query6[A, B, C, D, E, F]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -149,12 +304,29 @@ func (q *Query6[A, B, C, D, E, F]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query6.Next].
+func (q *Query6[A, B, C, D, E, F]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query6.Next].
 func (q *Query6[A, B, C, D, E, F]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query6.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query6[A, B, C, D, E, F]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query6.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query6[A, B, C, D, E, F]) Get() (*A, *B, *C, *D, *E, *F) {
@@ -167,7 +339,20 @@ func (q *Query6[A, B, C, D, E, F]) Get() (*A, *B, *C, *D, *E, *F) {
 		(*F)(unsafe.Add(q.columnPtrF, index*q.itemSizeF))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query6.NextTable].
+func (q *Query6[A, B, C, D, E, F]) GetColumns() ([]A, []B, []C, []D, []E, []F) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len],
+		q.columnD.data.Interface().([]D)[:q.table.len:q.table.len],
+		q.columnE.data.Interface().([]E)[:q.table.len:q.table.len],
+		q.columnF.data.Interface().([]F)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query7.NextTable].
 func (q *Query7[A, B, C, D, E, F, G]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -176,12 +361,29 @@ func (q *Query7[A, B, C, D, E, F, G]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query7.Next].
+func (q *Query7[A, B, C, D, E, F, G]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query7.Next].
 func (q *Query7[A, B, C, D, E, F, G]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query7.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query7[A, B, C, D, E, F, G]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query7.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query7[A, B, C, D, E, F, G]) Get() (*A, *B, *C, *D, *E, *F, *G) {
@@ -195,7 +397,21 @@ func (q *Query7[A, B, C, D, E, F, G]) Get() (*A, *B, *C, *D, *E, *F, *G) {
 		(*G)(unsafe.Add(q.columnPtrG, index*q.itemSizeG))
 }
 
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query7.NextTable].
+func (q *Query7[A, B, C, D, E, F, G]) GetColumns() ([]A, []B, []C, []D, []E, []F, []G) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len],
+		q.columnD.data.Interface().([]D)[:q.table.len:q.table.len],
+		q.columnE.data.Interface().([]E)[:q.table.len:q.table.len],
+		q.columnF.data.Interface().([]F)[:q.table.len:q.table.len],
+		q.columnG.data.Interface().([]G)[:q.table.len:q.table.len]
+}
+
 // Next advances the query's cursor to the next entity.
+//
+// For alternative, faster iteration over tables, use [Query8.NextTable].
 func (q *Query8[A, B, C, D, E, F, G, H]) Next() bool {
 	if int64(q.cursor.index) < q.cursor.maxIndex {
 		q.cursor.index++
@@ -204,12 +420,29 @@ func (q *Query8[A, B, C, D, E, F, G, H]) Next() bool {
 	return q.nextTableOrArchetype()
 }
 
+// NextTable advances the query's cursor to the next table.
+//
+// For alternative iteration over entities, use [Query8.Next].
+func (q *Query8[A, B, C, D, E, F, G, H]) NextTable() bool {
+	return q.nextTableOrArchetype()
+}
+
 // Entity returns the current entity.
+// Use this with entity iteration using [Query8.Next].
 func (q *Query8[A, B, C, D, E, F, G, H]) Entity() Entity {
 	return q.table.GetEntity(q.cursor.index)
 }
 
+// Entities returns the entities of the current table.
+// Use this with table-based iteration using [Query8.NextTable].
+//
+// ⚠️ Do not set/replace any of the slice elements!
+func (q *Query8[A, B, C, D, E, F, G, H]) Entities() []Entity {
+	return q.table.entities.data.Interface().([]Entity)[:q.table.len:q.table.len]
+}
+
 // Get returns the queried components of the current entity.
+// Use this with entity iteration using [Query8.Next].
 //
 // ⚠️ Do not store the obtained pointers outside of the current context (i.e. the query loop)!
 func (q *Query8[A, B, C, D, E, F, G, H]) Get() (*A, *B, *C, *D, *E, *F, *G, *H) {
@@ -222,4 +455,17 @@ func (q *Query8[A, B, C, D, E, F, G, H]) Get() (*A, *B, *C, *D, *E, *F, *G, *H) 
 		(*F)(unsafe.Add(q.columnPtrF, index*q.itemSizeF)),
 		(*G)(unsafe.Add(q.columnPtrG, index*q.itemSizeG)),
 		(*H)(unsafe.Add(q.columnPtrH, index*q.itemSizeH))
+}
+
+// GetColumns returns the queried component columns of the current table.
+// Use this with table-based iteration using [Query8.NextTable].
+func (q *Query8[A, B, C, D, E, F, G, H]) GetColumns() ([]A, []B, []C, []D, []E, []F, []G, []H) {
+	return q.columnA.data.Interface().([]A)[:q.table.len:q.table.len],
+		q.columnB.data.Interface().([]B)[:q.table.len:q.table.len],
+		q.columnC.data.Interface().([]C)[:q.table.len:q.table.len],
+		q.columnD.data.Interface().([]D)[:q.table.len:q.table.len],
+		q.columnE.data.Interface().([]E)[:q.table.len:q.table.len],
+		q.columnF.data.Interface().([]F)[:q.table.len:q.table.len],
+		q.columnG.data.Interface().([]G)[:q.table.len:q.table.len],
+		q.columnH.data.Interface().([]H)[:q.table.len:q.table.len]
 }